Move ComponentManager
