+++ src/main/java/seedu/address/model/ModelManager.java
@@ -4,7 +4,7 @@ import javafx.collections.transformation.FilteredList;
 import seedu.address.commons.core.LoggerManager;
 import seedu.address.commons.util.StringUtil;
 import seedu.address.events.model.LocalModelChangedEvent;
-import seedu.address.main.ComponentManager;
+import seedu.address.commons.core.ComponentManager;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.ReadOnlyPerson;
 import seedu.address.model.person.UniquePersonList;
